+++ cypress/e2e/menus/print.cy.ts
@@ -4,7 +4,7 @@ describe('菜单-打印', () => {
   beforeEach(() => {
     cy.visit('http://localhost:3000/canvas-editor/')
 
-    cy.get('canvas').should('have.length', 2)
+    cy.get('.ce-page-container canvas').should('have.length', 2)
   })
 
   it('打印', () => {