+++ src/editor/core/event/handlers/keydown/left.ts
@@ -7,7 +7,7 @@ import {
   getIsBlockElement,
   getNonHideElementIndex
 } from '../../../../utils/element'
-import { isMod } from '../../../../utils/hotkey'
+import { isApple } from '../../../../utils/ua'
 import { CanvasEvent } from '../../CanvasEvent'
 
 export function left(evt: KeyboardEvent, host: CanvasEvent) {
@@ -39,7 +39,7 @@ export function left(evt: KeyboardEvent, host: CanvasEvent) {
   }
   // 单词整体移动
   let moveCount = 1
-  if (isMod(evt)) {
+  if (isApple ? evt.altKey : evt.ctrlKey) {
     const LETTER_REG = draw.getLetterReg()
     // 起始位置
     const moveStartIndex =