+++ src/editor/core/worker/WorkerManager.ts
@@ -93,4 +93,11 @@ export class WorkerManager {
       })
     })
   }
+
+  public destroy() {
+    this.wordCountWorker.terminate()
+    this.catalogWorker.terminate()
+    this.groupWorker.terminate()
+    this.valueWorker.terminate()
+  }
 }