+++ src/editor/dataset/constant/Editor.ts
@@ -8,7 +8,8 @@ export const EDITOR_CLIPBOARD = `${EDITOR_PREFIX}-clipboard`
 export const defaultModeRuleOption: Readonly<DeepRequired<IModeRule>> = {
   print: {
     imagePreviewerDisabled: false,
-    backgroundDisabled: false
+    backgroundDisabled: false,
+    filterEmptyControl: true
   },
   readonly: {
     imagePreviewerDisabled: false