+++ src/editor/dataset/constant/Magnifier.ts
@@ -1,0 +1,8 @@
+import { IMagnifierOption } from '../../interface/Magnifier'
+
+export const defaultMagnifierOption: Readonly<Required<IMagnifierOption>> = {
+  disabled: true,
+  size: 120,
+  zoom: 2,
+  borderColor: '#efefef'
+}