+++ src/editor/dataset/constant/Watermark.ts
@@ -1,6 +1,6 @@
 import { IWatermark } from '../../interface/Watermark'
 import { NumberType } from '../enum/Common'
-import { WatermarkType } from '../enum/Watermark'
+import { WatermarkLayer, WatermarkType } from '../enum/Watermark'
 
 export const defaultWatermarkOption: Readonly<Required<IWatermark>> = {
   data: '',
@@ -13,5 +13,6 @@ export const defaultWatermarkOption: Readonly<Required<IWatermark>> = {
   font: 'Microsoft YaHei',
   repeat: false,
   gap: [10, 10],
-  numberType: NumberType.ARABIC
+  numberType: NumberType.ARABIC,
+  layer: WatermarkLayer.BOTTOM
 }