+++ src/editor/interface/Magnifier.ts
@@ -1,0 +1,6 @@
+export interface IMagnifierOption {
+  disabled?: boolean
+  size?: number
+  zoom?: number
+  borderColor?: string
+}