+++ src/editor/utils/element.ts
@@ -145,7 +145,7 @@ export function formatElementList(
       })
       // 追加节点
       if (valueList.length) {
-        const listId = getUUID()
+        const listId = el.listId || getUUID()
         for (let v = 0; v < valueList.length; v++) {
           const value = valueList[v]
           value.listId = listId
@@ -157,7 +157,9 @@ export function formatElementList(
         // 尾部如果不是换行符则补充一个换行符
         if (
           elementList[i] &&
-          !START_LINE_BREAK_REG.test(elementList[i].value)
+          (elementList[i].valueList?.length
+            ? !START_LINE_BREAK_REG.test(elementList[i].valueList![0].value)
+            : !START_LINE_BREAK_REG.test(elementList[i].value))
         ) {
           elementList.splice(i, 0, {
             value: ZERO